file: library: fix ID generation for 32bits env

+++ src/modules/file/melo_library_file.c
@@ -121,9 +121,9 @@ melo_library_file_get_info (MeloBrowser *browser)
 }
 
 static inline gchar *
-melo_library_gen_id (gint64 id)
+melo_library_gen_id (gint id)
 {
-  return g_strdup_printf ("%ld", id);
+  return g_strdup_printf ("%d", id);
 }
 
 static gboolean
